+++ src/Exception/NoRouteProvided.php
@@ -3,6 +3,9 @@ declare(strict_types = 1);
 
 namespace Innmind\Router\Exception;
 
-class LogicException extends \LogicException implements Exception
+/**
+ * @internal
+ */
+final class NoRouteProvided extends \LogicException
 {
 }