+++ src/Exception/NotFound.php
@@ -3,6 +3,9 @@ declare(strict_types = 1);
 
 namespace Innmind\Router\Exception;
 
-class DomainException extends \DomainException implements Exception
+/**
+ * @internal
+ */
+final class NotFound extends \RuntimeException
 {
 }